# doc/produced/swagger/organisations.yaml
swagger: '2.0'
info:
  version: 1.0-SNAPSHOT
  title: USM-administration
basePath: /rest
paths:
...
  /organisations/{organisationId}:
    get:
      description: Retrieves a organisation according to the provided organisationId
      operationId: getOrganisation
      produces:
        - application/json
      parameters:
        - name: authorisation
          description: the JWT token authenticating the service requester
          in: header
          required: true
          type: string
        - name: organisationId
          in: path
          description: The requested organisation id
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: organisations response
          schema:
              $ref: '#/definitions/organisation'
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/errorModel'
    delete:
      description: Deletes an existing organisation
      operationId: deleteOrganisation
      produces:
        - application/json
      parameters:
        - name: authorisation
          description: the JWT token authenticating the service requester
          in: header
          required: true
          type: string
        - name: organisationId
          in: path
          description: The requested organisation id
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: response of the operation
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/errorModel'
...
  /organisations/endpoint/{endPointId}:
    get:
      description: Retrieves an end point according to the provided endPointId
      operationId: getEndPoint
      produces:
        - application/json
      parameters:
        - name: authorisation
          description: the JWT token authenticating the service requester
          in: header
          required: true
          type: string
        - name: endPointId
          in: path
          description: the requested endpoint id
          required: true
          type: integer
          format: int64
      responses:
        '200':
          description: endpoint response
          schema:
              $ref: '#/definitions/endpoint'
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/errorModel'
  /endpoint:
    post:
     description: Creates a new endpoint
     operationId: createEndpoint
     produces:
        - application/json
     parameters:
        - name: authorisation
          description: the JWT token authenticating the service requester
          in: header
          required: true
          type: string
        - name: endpoint
          in: body
          description: the new endpoint to be created
          required: true
          schema:
            $ref: '#/definitions/endpoint'
     responses:
         '200':
           description: endpoint response
           schema:
             $ref: '#/definitions/endpoint'
         default:
          description: unexpected error
          schema:
            $ref: '#/definitions/errorModel'
    put:
     description: Updates an existing endpoint
     operationId: updateEndpoint
     produces:
        - application/json
     parameters:
        - name: authorisation
          description: the JWT token authenticating the service requester
          in: header
          required: true
          type: string
        - name: endpoint
          in: body
          description: The endpoint to be updated
          required: true
          schema:
            $ref: '#/definitions/endpoint'
     responses:
         '200':
           description: endpoint response
           schema:
             $ref: '#/definitions/endpoint'
         default:
          description: unexpected error
          schema:
            $ref: '#/definitions/errorModel'
  /endpoint/{endpointId}:
    delete:
      description: Deletes an existing endpoint
      operationId: deleteEndpoint
      produces:
        - application/json
      parameters:
        - name: authorisation
          description: the JWT token authenticating the service requester
          in: header
          required: true
          type: string
        - name: endpointId
          in: path
          description: The requested endpointId id
          required: true
          type: string
      responses:
        '200':
          description: response of the operation
        default:
          description: unexpected error
          schema:
            $ref: '#/definitions/errorModel'              
...
definitions:
  organisation:
    required:
      - organisationId
      - name
      - status
      - nation
    properties:
      organisationId:
        type: integer
        format: int64
      name:
        type: string
      description:
        type: string
      nation:
        type: string
      status:
        type: string
      parent:
        type: string
      email:
        type: string
      assignedUsers:
        type: integer
        format: int32
      endpoints:
        type: array
        items:
          $ref: '#/definitions/endpoint'
        description: The list of endpoints associated with this specific organisation
  endpoint:
    required:
      - endpointId
      - name
      - status
      - URI
    properties:
      endpointId:
        type: integer
        format: int64
      name:
        type: string
      description:
        type: string
      URI:
        type: string
      status:
        type: string
      organisationName:
        type: string
      channels:
        type: array
        items:
          $ref: '#/definitions/channel'
        description: The list of channels associated with this specific endpoint
      endpointcontacts:
        type: array
        items:
          $ref: '#/definitions/endpointcontact'
        description: The list of endpointcontacts associated with this specific endpoint   
  channel:
    required:
      - channelId
      - dataflow
      - service
      - priority
    properties:
      channelId:
        type: integer
        format: int64
      dataflow:
        type: string
      service:
        type: string
      priority:
        type: string
      endpointId:
        type: integer
        format: int64
  endpointcontact:
    required:
      - endPointContactId
      - personId
    properties:
      endPointContactId:
        type: integer
        format: int64
      firstName:
        type: string
      lastName:
        type: string
      phoneNumber:
        type: string
      mobileNumber:
        type: string
      faxNumber:
        type: string
      email:
        type: string        
      endpointId:
        type: integer
        format: int64
      personId:
        type: integer
        format: int64
  errorModel:
    required:
      - code
      - message
    properties:
      code:
        type: integer
        format: int32
      message:
        type: string